requirements.txt
appdirs==1.4.4
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
dj-database-url==1.0.0
Django==3.2.16
djangorestframework==3.14.0
exceptiongroup==1.0.0rc9
gunicorn==20.1.0
h11==0.14.0
idna==3.4
lxml==4.9.1
outcome==1.2.0
packaging==21.3
psycopg2==2.9.5
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.4.0
urllib3==1.26.12
webdriver-manager==3.8.4
webdrivermanager==0.10.0
whitenoise==6.2.0
wsproto==1.2.0
